move some code out of webui.py, better frame chooser ui
