+++ webui.css
@@ -13,6 +13,10 @@
     filter: grayscale(100%);
 }
 
+#mainimage {
+    filter: drop-shadow(0px 0px 15px orangered);
+}
+
 #redbutton {
     background: orangered;
 }
